pyproject.toml
[tool.poetry]
name = "dydown"
version = "1.0.0"
description = "抖音视频下载器"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dydown"}]

[tool.poetry.dependencies]
python = "^3.8"
PyQt6 = "^6.6.1"
PyQt6-WebEngine = "^6.6.0"
requests = "^2.31.0"
yt-dlp = "^2023.12.30"
ffmpeg-python = "^0.2.0"
pystray = "^0.19.5"
Pillow = "^10.1.0"
qrcode = "^7.4.2"
packaging = "^23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
pyinstaller = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dydown = "dydown.main:main" 